Finish Grabbing Android Device Token for React Native
