+++ android/app/src/main/java/com/michaelgill1969/checkin/FCMPackage.java
@@ -1,4 +1,4 @@
-// CustomToastPackage.java
+// FCMPackage.java
 
 package com.michaelgill1969.checkin;
 
@@ -11,19 +11,22 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class CustomToastPackage implements ReactPackage {
+public class FCMPackage implements ReactPackage {
 
   @Override
-  public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
+  public List<ViewManager> createViewManagers(
+    ReactApplicationContext reactContext
+  ) {
     return Collections.emptyList();
   }
 
   @Override
   public List<NativeModule> createNativeModules(
-                              ReactApplicationContext reactContext) {
+    ReactApplicationContext reactContext
+  ) {
     List<NativeModule> modules = new ArrayList<>();
 
-    modules.add(new ToastModule(reactContext));
+    modules.add(new FCMModule(reactContext));
 
     return modules;
   }
