+++ android/app/src/main/java/com/michaelgill1969/checkin/MainActivity.java
@@ -1,13 +1,6 @@
 package com.michaelgill1969.checkin;
 
 import com.facebook.react.ReactActivity;
-import android.os.Bundle;
-import androidx.annotation.NonNull;
-
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.iid.InstanceIdResult;
 
 public class MainActivity extends ReactActivity {
 
@@ -19,24 +12,4 @@ public class MainActivity extends ReactActivity {
     protected String getMainComponentName() {
         return "checkin";
     }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        FirebaseInstanceId.getInstance().getInstanceId()
-            .addOnCompleteListener(new OnCompleteListener<InstanceIdResult>() {
-                @Override
-                public void onComplete(@NonNull Task<InstanceIdResult> task) {
-                    if (!task.isSuccessful()) {
-                        return;
-                    }
-
-                    // Get new Instance ID token
-                    String token = task.getResult().getToken();
-                }
-            });
-        // TODO: Monitor Token Generation
-        // https://firebase.google.com/docs/cloud-messaging/android/client#monitor-token-generation
-  }
 }
