+++ android/app/src/main/java/com/michaelgill1969/checkin/MainApplication.java
@@ -7,7 +7,7 @@ import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.michaelgill1969.checkin.CustomToastPackage;
+import com.michaelgill1969.checkin.FCMPackage;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
@@ -26,7 +26,7 @@ public class MainApplication extends Application implements ReactApplication {
           List<ReactPackage> packages = new PackageList(this).getPackages();
           // Packages that cannot be autolinked yet can be added manually here, for example:
           // packages.add(new MyReactNativePackage());
-          packages.add(new CustomToastPackage());
+          packages.add(new FCMPackage());
           return packages;
         }
